Update parafit_script.R
Changed read.nexus fxns to readNexus (phytools) because readNexus stores the node label bootstrap values. If the nexus file is in "raxml" format, the node labels are represented as [&label=bs_value] in the nexus file. These labels will be stored as the node.labels in the phylo objects, and can later be added into the final plot.
Added options to modify tip.labels (the taxa names) using str_replace fxn; this is useful if you need to replace a prefix/suffix/punctuation on tip labels (especially things like '-' which R sometimes doesn't like.
Added lines to include node labels on final cophylogeny plot in PDF (to avoid having to add them in manually on the final PDF).